Add ssh key management to ssh-proxy command
+ Parse the public key from the IdentityFile that is passed as an
argument to the ssh-proxy command and send a new websocket message
containing this public key in order to have bastion setup this key as an
authorized key on the target.

+ If the identity file passed to ssh-proxy is the ssh key path that is
managed through thoum then generate a new ephemeral ssh key and
overwrite this file each time a new ssh tunnel is created.