Remove shell reattach logic in zli (#94)
Handle shell unattached events by reporting an error to the user and
exiting the zli without closing the connection. This will allow a zli
session to be accessed from the web app without closing the
connection. Once opened in the web app this connection cannot be
currently re-attached to in the zli, however a future enhancement could
expose active connections in the zli-space to allow such a feature.