(CWC-599) Feat: Keysplitting for Interactive Shell (#89)
* Add removeKeysplittingData method to config.service

* Add Keysplitting protocol for interactive ssm shells

* Add agent version check

Fallback to using sshShellWebsocketService for ssm targets with agent
versions that are not compatible with keysplitting. The
sshShellWebsocketService is backwards compatible with these ssm targets.

* Add refresh OAuth id token functionality (#93)

* Add refresh OAuth id token functionality while in the middle of a session

* Add shell unattach/reattach logic to terminal.ts

Handle shell unattached events by reporting an error to the user and
exiting the zli without closing the connection. This will allow a zli
session to be accessed from the web app without closing the
connection. Once opened in the web app this connection cannot be
currently re-attached to in the zli, however a future enhancement could
expose active connections in the zli-space to allow such a feature.

* Send terminal resize input when receiving a start shell event

+ This ensures the initial terminal dimensions are set in the zli for
keysplitting-enabled shells that must send terminal size messages as
keysplitting data messages.

* Update openid-client to 4.7.2

* Await on all occurences of cleanExit

* [Version] 4.14.0 -> 4.15.0

Co-authored-by: Thanasis Filippidis <aflpd@bu.edu>