Feat/quickstart ssh (#166)
Short summary: Adds a new command to the zli: zli quickstart. Quickstart aims to be an interactive onboarding experience for users accustomed to SSH. By the end of a successful quickstart execution, the user should have their selected SSH hosts fully integrated with BastionZero; this includes the creation of policies and environments.

How does it work?: Quickstart scans the user's ssh_config(5) file in order to find a list of potentially SSH-able hosts in which the BastionZero autodiscovery script can be executed on; the config file contains the list of parameters and authorization method (SSH key) to use when quickstart begins its SSH connection. After validating the parsed config file, quickstart SSHes to a list of hosts chosen by the user. The autodiscovery script is executed on each host concurrently, and quickstart polls the Bastion waiting for each target to come online. In addition, quickstart creates the necessary policies and environments, so that the user running quickstart can connect to their targets once they're fully registered. A new policy+environment is created (if they don't already exist) for each unique SSH username parsed from the ssh_config file.

Co-authored-by: Sebby <lipmas93@gmail.com>