Add access token header in groups/list endpoint for microsoft users (#224)
* Add access token header in groups/list endpoint for microsoft users

This is needed in order for the microft groups integration to work
properly.  The backend needs an access token with the required scopes
for fetching group information from microsoft graph API. This only
applies to users who have gone through the azure groups integration flow
in the webapp and whose access token contains the scopes User.Read.All
and Group.Read.All. If the access token does not contain these scops we
can skip passing the token to the backend and the groups/list endpoint
will simply return an empty list of groups.