Quickstart refresh (#232)
* Missing port is no longer a parse error. Default to port 22

* Remove prompt to handle invalid SSH config file

* Remove quickstart functionality to connect to shell at end

* Add transcript feature

+ Create new service consoleWithTranscript service

* Add progress bar, tips, and steps

* Capture SIGINT even when running the autodiscovery script

* Display visual error when env/policy creation fails

* Add login to quickstart flow

* Clear screen while preserving scrollback

+ Don't assume err is of type Error in fail()

Issue: reject() in login() will log "Login timeout reached" even if
resolve() is called.

Solution: Changes the setTimeout case to call reject() with an error
message instead of executing a side effect where the logger is called.
This refactor is needed because login is now used in another place
(quickstart) where execution continues--that is, there is more todo
after logging in during quickstart.

The reason we never saw this issue earlier is because login would
immediately cleanExit after it was done successfully.

In JS, reject() can be called more than once, but the first call to
resolve() or reject() is the final state of the Promise. However, if the
call to resolve() or reject() executes a side effect (e.g. log a
message), this action still occurs even if the promise is resolved.
Another way to fix this is to cancel the timeout when resolve is called.

* Bump version to 5.1.39