Fixes quickstart
Due to the removal of manual name scheme, an assumption quickstart makes
at the end of registration is incorrect. Previously, quickstart would
look up the ssh host name in the config to find some details it needs to
create policies and print the final message. This lookup depends on the
ssm target name matching the name of the SSH host block in the user's
config.

Due to not using manual name anymore, we can't make this assumption.
Furthermore, this was probably bad design in the first place. Quickstart
shouldn't have to look in its dictionary again to find ssh host details;
it should know it at the time of successful registration. This commit
does exactly that; after registration is complete (autodiscovery
complete), the method will return a composite type that includes both
the ssh config of the registered host and the ssm target summary
details.