Update terminal service to work with geo-latency CS (#279)
* Update terminal service to work with geo-latency CS

* Minor version bump

* Update webshell-common-ts

* Return region in list-targets command (#292)

* System test for geo-ip connection service - SSM targets (#295)

* Fix race condition in connect suite

The connect test previously used a spy to get the
terminalRunningObservable that shell-utils.ts uses. Even though this is
the same observable, the Rx subscription that the test makes to wait for
the ShellReady event to be sent is a separate subscription.

Race condition: By the time the test creates its subscription, the
start() request has already been sent / websocket is already
established. Therefore, there is a chance for the client to receive
ShellStart and for the test to miss this critical event (i.e. we subbed
to late).

If the test misses this event, then the terminalIsReadyForInput promise
never resolves. Jest kills the test eventually, but now we failed due to
having a race condition outside of the SUT (in the test itself) which
makes the system test not so good.

----

The test has been simplified to not rely on terminalRunning spy.
Instead, we just keep sending input until receiving the expected output
(this is done with a 30sec timeout).

However, this change is not ideal. It would be nice if we didn't have to
resend input / give the test a signal that input is now able to be sent.
A future PR will attempt to remedy this, fixing something for both the
system test and the zli shell-utils itself. Note: Please keep in mind
there exist other race conditions in the connect flow that are not
caused by system test (e.g. handler shell not ready / pty not ready /
not receiving ack in ssm data channel). These other race conditions can
be fixed in new agent + backend changes.

+ Fix race condition in shell-utils as well. Technically, we could also
  miss events in regular usage. This is easier to fix as we have more
  control on order of operations in the client code than the system test
  code. Fixed this by subscribing to the terminalRunning observable
  prior to calling `start()`

* Add geo-ip connection service test assertion

The connect suite now checks that the ShellConnectionAuthDetails
response matches an expected AWS region. This enhances the connect suite
against our new regional connection-service because in addition to
typing a command and receiving expected output, this extra check ensures
we connected to a connection node in an expected region based on the
droplet's region.

+ Add EXTRA_REGIONS environment variable. System test CodeBuild job can
  be configured with different set of regions to test. Each region in
  this list corresponds to a new droplet to test in the connect suite.

  By default this list contains "ap-northeast-1" (Tokyo) if the env-var
  is not set.

  The system test will fail immediately if there is no known conversion
  from AWS region to DigitalOcean region.

+ Add region variable in SsmTestTarget type, so that createDroplet()'s
  region is configurable.

* Add unique id to droplet name

* Test package manager path with geo-ip

* Fix lint errors