CWC-1507 + System test improvements (#316)
* Pass target/agent's base64 encoded public key to daemon

This public key is received from the Bastion. The daemon needs this
pubkey to verify received KS messages from the agent. It also commits to
it in the SYN.

* Pass --agentVersion flag to daemon from target details

* System test improvements

+ Update AmazonLinux2 and BzeroVTAL2TestImage DigitalOcean images. git
  was broken in the old one
+ Add SYSTEM_TEST_TAGS environment variable. Optionally set this variable to
  comma-separted list of tags to override default tag of "system-tests"
+ Add BZERO_AGENT_BRANCH environment variable. Optionally set this
  variable to compile the bzero-agent from a specific github.com/bastionzero/bzero remote
  branch when running the VT tests

* Use dots-wrapper typescript library instead of digitalocean-js

This library is more up to date to DigitalOcean's APIs and has more support

* Add ability to test specific kube agent image hosted @ DigitalOcean

Use new env var: BZERO_KUBE_AGENT_IMAGE. <image-name>@<image-version>

The image must exist in our private DigitalOcean container registry

* Bump to version 5.5.5

* Use registration key when doing helm install