CWC-1518: Generate SSH config and support case-sensitive targets (#329)
To ease the transition of our end users and our end users' customers to BastionZero, we will generate the additional configuration setup needed to add to an individual’s ssh config. This serves to onboard a new BastionZero user faster and also prevents small user errors that can lead to frustration.

This PR adds the following command to the ZLI:

zli generate sshConfig
Which will do two things:

Populate a new .ssh/bz-config file to assist the user in connecting to BZ targets
If that file is not already Included in the user's primary .ssh/config file, we add a line to include it
Importantly, there is also a small change in how the ProxyCommand gets formed, both by zli generate sshConfig and zli ssh-proxy-config -- instead of using the %h token, which SSH converts to lowercase, we now use the $n token, which is left unaltered from what is entered in the original ssh command. This will allow even mixed-case target names to be discovered by an ssh proxy through ZLI